[ISSUE-261] Refine message class hierarchy
