+++ common/src/main/java/com/aliyun/emr/rss/common/network/protocol/ChunkFetchSuccess.java
@@ -30,7 +30,7 @@ import com.aliyun.emr.rss.common.network.buffer.NettyManagedBuffer;
  * may be written by Netty in a more efficient manner (i.e., zero-copy write).
  * Similarly, the client-side decoding will reuse the Netty ByteBuf as the buffer.
  */
-public final class ChunkFetchSuccess extends AbstractResponseMessage {
+public final class ChunkFetchSuccess extends ResponseMessage {
   public final StreamChunkSlice streamChunkSlice;
 
   public ChunkFetchSuccess(StreamChunkSlice streamChunkSlice, ManagedBuffer buffer) {
