+++ common/src/main/java/com/aliyun/emr/rss/common/network/protocol/MessageDecoder.java
@@ -42,43 +42,9 @@ public final class MessageDecoder extends MessageToMessageDecoder<ByteBuf> {
   @Override
   public void decode(ChannelHandlerContext ctx, ByteBuf in, List<Object> out) {
     Message.Type msgType = Message.Type.decode(in);
-    Message decoded = decode(msgType, in);
+    Message decoded = Message.decode(msgType, in);
     assert decoded.type() == msgType;
     logger.trace("Received message {}: {}", msgType, decoded);
     out.add(decoded);
-  }
-
-  private Message decode(Message.Type msgType, ByteBuf in) {
-    switch (msgType) {
-      case ChunkFetchRequest:
-        return ChunkFetchRequest.decode(in);
-
-      case ChunkFetchSuccess:
-        return ChunkFetchSuccess.decode(in);
-
-      case ChunkFetchFailure:
-        return ChunkFetchFailure.decode(in);
-
-      case RpcRequest:
-        return RpcRequest.decode(in);
-
-      case RpcResponse:
-        return RpcResponse.decode(in);
-
-      case RpcFailure:
-        return RpcFailure.decode(in);
-
-      case OneWayMessage:
-        return OneWayMessage.decode(in);
-
-      case PushData:
-        return PushData.decode(in);
-
-      case PushMergedData:
-        return PushMergedData.decode(in);
-
-      default:
-        throw new IllegalArgumentException("Unexpected message type: " + msgType);
-    }
   }
 }
